Debug mode (#5)
* Add debug mode to config

Add a debug mode entry (key `DEBUG_MODE`) in the default config. This
entry should only be set to either `true` or `false` so it could be
parsed as boolean.

This entry may be used to enable debug output, as an example.

* Update log flags with debug mode in mind

Only set the `log.LlongFile` flag if the debug mode entry in the config
is set to `true`.

This avoids cluttering up production logs, while allowing devs to
quickly be able to find log places in while working.